+++ website/app/docs/concepts/consumers/page.tsx
@@ -1,0 +1,6 @@
+import MarkdownPage from '../../../components/MarkdownPage';
+import Content from '../../../../content/concepts/consumers.mdx';
+
+export default function Consumers() {
+  return <MarkdownPage><Content /></MarkdownPage>;
+}