+++ website/app/docs/concepts/message-intent/page.tsx
@@ -1,0 +1,6 @@
+import MarkdownPage from '../../../components/MarkdownPage';
+import Content from '../../../../content/concepts/message-intent.mdx';
+
+export default function MessageIntent() {
+  return <MarkdownPage><Content /></MarkdownPage>;
+}