+++ website/app/docs/concepts/messages/page.tsx
@@ -1,0 +1,6 @@
+import MarkdownPage from '../../../components/MarkdownPage';
+import Content from '../../../../content/concepts/messages.mdx';
+
+export default function Messages() {
+  return <MarkdownPage><Content /></MarkdownPage>;
+}