+++ website/app/docs/concepts/receive-endpoints/page.tsx
@@ -1,0 +1,6 @@
+import MarkdownPage from '../../../components/MarkdownPage';
+import Content from '../../../../content/concepts/receive-endpoints.mdx';
+
+export default function ReceiveEndpoints() {
+  return <MarkdownPage><Content /></MarkdownPage>;
+}