+++ website/app/docs/concepts/reliability/page.tsx
@@ -1,0 +1,6 @@
+import MarkdownPage from '../../../components/MarkdownPage';
+import Content from '../../../../content/concepts/reliability.mdx';
+
+export default function Reliability() {
+  return <MarkdownPage><Content /></MarkdownPage>;
+}