+++ website/app/docs/concepts/requests/page.tsx
@@ -1,0 +1,6 @@
+import MarkdownPage from '../../../components/MarkdownPage';
+import Content from '../../../../content/concepts/requests.mdx';
+
+export default function Requests() {
+  return <MarkdownPage><Content /></MarkdownPage>;
+}