+++ website/app/docs/concepts/routing-topology/page.tsx
@@ -1,0 +1,6 @@
+import MarkdownPage from '../../../components/MarkdownPage';
+import Content from '../../../../content/concepts/routing-topology.mdx';
+
+export default function RoutingTopology() {
+  return <MarkdownPage><Content /></MarkdownPage>;
+}